requirements.txt
PyYAML >= 3.12
selenium >= 3.11.0
requests >= 2.18.4